pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.0",
    "uvicorn>=0.41.0",
]
